add reg class name info

+++ include/cppbind/utility/EnumHelper.hpp
@@ -29,7 +29,7 @@ public:
    }
    std::string valueStr() const {
        int x = value;
-       assert(x < T::enum_str_info_len);
+       assert(x < (int)T::enum_str_info_len);
        return T::enum_str_info[x];
    }
 private:
